backend/functions/data-service/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.13.1
